ai/requirements.txt
opencv-python==4.8.1.78
requests==2.31.0
numpy==1.24.3
ultralytics==8.0.186
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0